Update script to handle requests that haven't been submitted yet
These requests do not have a DAR id assigned yet, so they were
breaking the previous assertion in the code that all requests had
both a DAR and an abbreviation. Separate out the DAR block parsing
for each phs into its own function, and add a main function. This
way, the code can be imported into python without running everything.

Note that it would be nice to further functionize this, or define
classes with methods that the main function can then call. Also
would be good to add some tests.